Back tracked like agreed, and approached this a different way.
We now have a spellbook that acts as the processor for aknowledging whether the player has put the correct combo in or not.

We then have a spell virtual class, that houses the individual spells as derived classes, the combo entered by the player then points to one of these and calls away for the spell to be cast.

System currently working well, just need to implement actual spell logic--currently the spells keep track of their own quantity of charges, and identify where the mouse click was to cast at a target location and print those details to console.

Project currently feels pretty solid, though haven't been doing enough commenting.