Got the fireballs correctly added to the vector, and have them drawing.
The circular dependancy issue ended up being hours and hours of solving, I just have no experience with them and did not know what I was dealing with.

Still not 100% sure how I beat it, but now have correct use of forward decleration VS includes.

Long story short, forward dec is fine for accessing some class or assigning a pointer before the class dec is finished----include is required if you want to access the class and any of its properties.

Probs merge this branch back in with main now that we have our functionality back.